Extract account module
